Render a table of contents in the sidebar
The tech-docs-template sidebar contains a table of content for the
current page. In Jekyll, the easiest way to do that is with [Kramdown's
table of contents
generation](https://kramdown.gettalong.org/converter/html.html#toc).

Unfortunately, that relies on the table of contents being in the same
Markdown block as the table itself, which we don't want. To get around
that, we're hackily dumping the raw Markdown page content into a
variable with the Kramdown table-of-contents [inline attribute
list](https://kramdown.gettalong.org/syntax.html#inline-attribute-lists).
Once we have that, we markdownify it, then replace the rendered page in
*that* output with the empty string. This also won't work if the
contents of… um… `content` contains Jekyll code that needs to be
executed, because `markdownify` doesn't do that. I'm not really sure
how to fix that yet.

This is a massive hack, but it's this or do it in JavaScript, or in a
Jekyll plugin (which won't work on GitHub Pages).

It also turns out that, because we're always going to have a table of
contents, we can hard-code the data-module on .app-pane__body to
`in-page-navigation`. (This appears to be based on the
tech-docs-template team's desire to allow the sidebar to be overridden.)