Ensure JavaScript works under Jekyll
The JavaScript in the tech-docs-template uses the Rails asset pipeline.
We don't have that in Jekyll, so we should replace the `require`
statements with Jekyll includes… but that requires that all the
JavaScript be in _includes, so we should move them there.

We also need to add yaml front-matter to the remaining JavaScript files
so Jekyll renders them correctly, and finally fix the JavaScript
include in the HTML.

```bash
mkdir _includes/javascripts
for source in javascripts/_*; do
  target=${source/javascripts\/_/_includes\/javascripts/}
  git mv $source $target
done
find javascripts _includes/javascripts -name '*.js' -print0 | \
  xargs -0 sed -i '' -E \
  -e 's/\/\/= require _([^$]+)/{% include javascripts\/\1.js %}/g' \
find javascripts -name '*.js' -print0 | \
  xargs -0 sed -i '' -E \
  -e '1s/^/---\
---\
/g'
find _layouts _includes -name '*.html' -print0 | \
  xargs -0 sed -i '' -E \
  -e 's/{{ javascript_include_tag :([^S]+) }}/<script src="{{ "\/javascripts\/\1.js" | relative_url }}" type="text\/javascript" charset="utf-8"><\/script>/g'
```