Make table of contents rendering more robust
If the page contains liquid markup, `markdownify` doesn't transform
that, meaning that the `replace` filter in the old code would fail and
the sidebar would end up containing the entire page again.

Because we know that the table of contents comes last in generated
content, we can split on some markup that precedes it and output that.

I think this will only break, now, if the headings themselves contain
(or are generated by) liquid code.