Describe this site, and how to contribute
We want to be clear about the purpose of this site, and how people can
add new guidance. This will be the site's index page (replacing the
readme), and should give readers everything they need to get where they
want to go.

In future, it'll also link to all the standards and guidance the site
will contain, but we have none yet.