Autogenerate the list of standards on the homepage
Rather than having to manually manage and update the list of standards
on the homepage, it would be nice if we can get Jekyll to do it for us.

To do that, we first mark everything in `/standards/` as a standard
from Jekyll's config. We can then filter `site.pages` to just get those
items, group them by their `category`, and render the list of standards
with appropriate headings.

Unfortunately, the headings generated here are rendered using Liquid
and so don't render correctly in the table of contents sidebar. This is
because we're unable (so far as I can tell) to parse the Liquid tags
before running `markdownify`. To avoid rendering garbage headings,
then, we apply the [`{:.no_toc}` Kramdown
attribute](https://kramdown.gettalong.org/converter/html.html#toc)
(which excludes them from) the table of contents.

This also adds an appropriate category to `storing-source-code.md`, and
adds an example to the index.