Add a Liquid-handling table of contents renderer
It turns out that @allejo, who I found on #jekyll in Freenode, has
written [a pure Liquid table of contents
renderer](https://allejo.io/blog/a-jekyll-toc-in-liquid-only/) that
works based on the final, rendered HTML rather than the Markdown. This
means it doesn't have problems with headers that are generated by
Liquid logic, and is thus more robust than my table of contents
generator. Given that, we can just integrate that table of contents
rendered.

This also adds the appropriate attribution for our use of @allejo's code.

The table of contents include does the following:

0. Accept, as input, the page content rendered as HTML (and, thus,
   already parsed by Liquid)
1. Split the HTML on `<h` and loop over the resulting `nodes` (making
   each `node` contain the remainder of the opening tag, and everything
   up to the next `<h`)
2. Generate new Markdown for each `node` by deconstructing the HTML up
   to the first `</h` (excluding headers and stripping markup according
   to input parameters)
5. Passes the rendered Markdown (in `my_toc`) through `markdownify` to
   output it.