Adds MoJ Development principles
I think it would be better to make our development principles more
visible for all to see and to have a clear understanding what our principles
are.

Both principles were originally set out in two separate google documents
and have not been reviewed or update for ~2years. This commit is merely
to port those two documents to this project.

I suspect that once we increase the visibility of these principles and
they have been reviewed that there will be further changes to them.