Add a standard for domain names for services
We have a variety of different approaches to domain names for the services
we've built. Not all services are on a *.gov.uk domain, and many services
which are on a *.gov.uk domain in production are using *.dsd.io domains for
their pre-production environments.

For (at least) the second time this year, we've found that several services on
*.dsd.io are being flagged by Chrome as potential phishing sites. We suspect
that that's because they look like GOV.UK services but aren't on a *.gov.uk
domain. We want to move those services to *.gov.uk domains to avoid this
issue, including some pre-production domains, and so need to define a standard
to help teams do that.