Fix the broken 'naming things' link in the website
In the 2nd paragraph of this page:

https://ministryofjustice.github.io/technical-guidance/standards/naming-domains/#naming-domains

...there is a link to "wider standards for naming"

This link works when running the site in local development mode (i.e.
via `jekyll serve`), but doesn't work when the site is published.

This commit applies a naïve fix, by replacing the broken relative link
with a working absolute link to the live site, however this is not
ideal. We need a broader fix that makes relative links compile correctly
when the site is published.