Replace content & config with tech-docs version
This change converts the site to use (a patched
version of) the gov.uk tech-docs template & gems.

https://github.com/alphagov/tech-docs-template

Details of the patch are here:
https://github.com/ministryofjustice/cloud-platform-user-guide/pull/393

This gives the site a search feature, as well as
dynamic review dates (i.e. you see a big red
banner at the foot of the page if the content was
not reviewed when it should have been).

Publishing is now done via a github action. See:

.github/workflows/publish.yml

...for details.